// src/types/typeId.ts
// https://reference.opcfoundation.org/v105/Core/docs/Part6/5.1.2/
export enum TypeId {
    Null = 0,
    Boolean = 1,
    SByte = 2,
    Byte = 3,
    Int16 = 4,
    Uint16 = 5,
    Int32 = 6,
    Uint32 = 7,
    Int64 = 8,
    Uint64 = 9,
    Float = 10,
    Double = 11,
    String = 12,
    DateTime = 13,
    GUID = 14,
    ByteString = 15,
    XMLElement = 16,
    NodeID = 17,
    ExpandedNodeID = 18,
    StatusCode = 19,
    QualifiedName = 20,
    LocalizedText = 21,
    ExtensionObject = 22,
    DataValue = 23,
    Variant = 24,
    DiagnosticInfo = 25,
}